Clear new hits

+++ src/graphic/text/font_io.cc
@@ -48,7 +48,7 @@ IFont* load_font(const std::string& face, int ptsize) {
 		throw BadFont("could not load font!: RWops Pointer invalid");
 	}
 
-	TTF_Font* font = TTF_OpenFontIndexRW(ops, true, ptsize, 0);
+	TTF_Font* font = TTF_OpenFontIndexRW(ops, static_cast<int>(true), ptsize, 0);
 	if (font == nullptr) {
 		throw BadFont(format("Font loading error for %s, %i pts: %s", face, ptsize, TTF_GetError()));
 	}
